+++ src/components/views/EmployeeViews.js
@@ -1,6 +1,7 @@
 import { Outlet, Route, Routes } from "react-router-dom"
 import { TicketForm } from "../tickets/TicketForm"
 import { TicketContainer } from "../tickets/TicketContainer"
+import { EmployeeList } from "../employees/employeeList"
 
 
 
@@ -18,6 +19,7 @@ export const EmployeeViews = () => {
 			}>
 
 				<Route path="tickets" element={ <TicketContainer />} />
+				<Route path="employees" element={ <EmployeeList />} />
 				
 			</Route>
 		</Routes>
